+++ app/javascript/components/Company/requests/insightsRequest.jsx
@@ -4,5 +4,8 @@ export const insightsRequest = async (companyUuid) => {
   const result = await apiRequest({
     url: `/api/frontend/companies/${companyUuid}/insights.json`,
   });
-  return result.insights.data.map((element) => element.attributes);
+  return {
+    data: result.insights.data.map((element) => element.attributes),
+    ratioType: result.insights.ratio_type
+  };
 };