+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - co-owners for accounts
 - expiration time for access tokens
 - creating/removing API access tokens
+- API documentation
+- fetching insights by API
 
 ### Modified
 - skip reseting invites email after accepting invite